more stuff

+++ UWUVCI AIO WPF/Classes/Injection.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
+using System.Linq;
 using System.Net;
 using System.Text;
 using System.Threading;
@@ -43,6 +44,77 @@ namespace UWUVCI_AIO_WPF
          * iniPath = Only used for N64. Path to the INI configuration. If "blank", a blank ini will be used.
          * darkRemoval = Only used for N64. Indicates whether the dark filter should be removed.
          */
+        static List<int> fiind(this byte[] buffer, byte[] pattern, int startIndex)
+        {
+            List<int> positions = new List<int>();
+            int i = Array.IndexOf<byte>(buffer, pattern[0], startIndex);
+            while (i >= 0 && i <= buffer.Length - pattern.Length)
+            {
+                byte[] segment = new byte[pattern.Length];
+                Buffer.BlockCopy(buffer, i, segment, 0, pattern.Length);
+                if (segment.SequenceEqual<byte>(pattern))
+                    positions.Add(i);
+                i = Array.IndexOf<byte>(buffer, pattern[0], i + 1);
+            }
+            return positions;
+        }
+        static void PokePatch(string rom)
+        {
+            byte[] search = { 0xD0, 0x88, 0x8D, 0x83, 0x42 };
+            byte[] test;
+            test = new byte[new FileInfo(rom).Length];
+            using (var fs = new FileStream(rom,
+                                 FileMode.Open,
+                                 FileAccess.ReadWrite))
+            {
+                try
+                {
+                    fs.Read(test, 0, test.Length - 1);
+
+                    var l = fiind(test, search, 0);
+                    byte[] check = new byte[4];
+                    fs.Seek(l[0] + 5, SeekOrigin.Begin);
+                    fs.Read(check, 0, 4);
+
+                    fs.Seek(0, SeekOrigin.Begin);
+                    if (check[3] != 0x24)
+                    {
+                        fs.Seek(l[0] + 5, SeekOrigin.Begin);
+                        fs.Write(new byte[] { 0x00, 0x00, 0x00, 0x00 }, 0, 4);
+
+                    }
+                    else
+                    {
+                        fs.Seek(l[0] + 5, SeekOrigin.Begin);
+                        fs.Write(new byte[] { 0x00, 0x00, 0x00 }, 0, 3);
+
+                    }
+                    check = new byte[4];
+                    fs.Seek(l[1] + 5, SeekOrigin.Begin);
+                    fs.Read(check, 0, 4);
+                    fs.Seek(0, SeekOrigin.Begin);
+                    if (check[3] != 0x24)
+                    {
+                        fs.Seek(l[1] + 5, SeekOrigin.Begin);
+                        fs.Write(new byte[] { 0x00, 0x00, 0x00, 0x00 }, 0, 4);
+
+                    }
+                    else
+                    {
+                        fs.Seek(l[1] + 5, SeekOrigin.Begin);
+                        fs.Write(new byte[] { 0x00, 0x00, 0x00 }, 0, 3);
+
+                    }
+                }
+                catch (Exception e)
+                {
+
+                }
+
+
+                fs.Close();
+            }
+        }
         private static string FormatBytes(long bytes)
         {
             string[] Suffix = { "B", "KB", "MB", "GB", "TB" };
@@ -1195,7 +1267,16 @@ namespace UWUVCI_AIO_WPF
                 delete = true;
                 mvvm.Progress = 40;
             }
-
+            if (mvvm.PokePatch)
+            {
+                mvvm.msg = "Applying PokePatch";
+                File.Copy(injectRomPath, Path.Combine(tempPath, "rom.gba"));
+                injectRomPath = Path.Combine(tempPath, "rom.gba");
+                PokePatch(injectRomPath);
+                delete = true;
+                mvvm.PokePatch = false;
+                mvvm.Progress = 40;
+            }
 
             using (Process psb = new Process())
             {
@@ -1212,7 +1293,7 @@ namespace UWUVCI_AIO_WPF
             if (delete)
             {
                 File.Delete(injectRomPath);
-                File.Delete(Path.Combine(toolsPath, "goombamenu.gba"));
+                if(File.Exists(Path.Combine(toolsPath, "goombamenu.gba")))File.Delete(Path.Combine(toolsPath, "goombamenu.gba"));
             }
         }
         private static void DownloadSysTitle(MainViewModel mvm)
